+++ src/main/java/com/structurizr/dsl/ComponentParser.java
@@ -58,4 +58,20 @@ final class ComponentParser extends AbstractParser {
         return component;
     }
 
+    void parseTechnology(ComponentDslContext context, Tokens tokens) {
+        int index = 1;
+
+        // technology <technology>
+        if (tokens.hasMoreThan(index)) {
+            throw new RuntimeException("Too many tokens, expected: technology <technology>");
+        }
+
+        if (!tokens.includes(index)) {
+            throw new RuntimeException("Expected: technology <technology>");
+        }
+
+        String technology = tokens.get(index);
+        context.getComponent().setTechnology(technology);
+    }
+
 }