+++ src/main/java/com/structurizr/dsl/ModelItemParser.java
@@ -1,6 +1,10 @@
 package com.structurizr.dsl;
 
+import com.structurizr.model.Element;
+
 final class ModelItemParser extends AbstractParser {
+
+    private final static int DESCRIPTION_INDEX = 1;
 
     private final static int TAGS_INDEX = 1;
 
@@ -22,6 +26,20 @@ final class ModelItemParser extends AbstractParser {
             String tags = tokens.get(i);
             context.getModelItem().addTags(tags.split(","));
         }
+    }
+
+    void parseDescription(ModelItemDslContext context, Tokens tokens) {
+        // description <description>
+        if (tokens.hasMoreThan(DESCRIPTION_INDEX)) {
+            throw new RuntimeException("Too many tokens, expected: description <description>");
+        }
+
+        if (!tokens.includes(DESCRIPTION_INDEX)) {
+            throw new RuntimeException("Expected: description <description>");
+        }
+
+        String description = tokens.get(DESCRIPTION_INDEX);
+        ((Element)context.getModelItem()).setDescription(description);
     }
 
     void parseUrl(ModelItemDslContext context, Tokens tokens) {