+++ docs/changelog.md
@@ -5,6 +5,8 @@
 - Adds the implied relationships functionality for custom elements.
 - "addDefaultElements" will now also add any connected custom elements.
 - Adds better support for custom elements when using element expressions.
+- Adds a `description` keyword for setting the description on elements.
+- Adds a `technology` keyword for setting the technology on containers, components, deployment nodes, and infrastructure nodes.
 
 ## 1.15.0 (2nd October 2021)